requirements.txt
google-api-python-client==2.108.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
flask==3.0.0